+++ XF/Mail/XF2/Mailer.php
@@ -7,6 +7,18 @@ use TickTackk\DeveloperTools\XF\Mail\XFCP_Mailer;
 
 class Mailer extends XFCP_Mailer
 {
+    /**
+     * @param \Swift_Mime_Message $message
+     * @param \Swift_Transport|null $transport
+     * @param array|null $queueEntry
+     * @param $allowRetry
+     *
+     * @return int
+     *
+     * @throws \XF\PrintableException
+     *
+     * @noinspection PhpHierarchyChecksInspection
+     */
     public function send
     (
         \Swift_Mime_Message $message,