+++ XF/Repository/CodeEvent.php
@@ -5,6 +5,8 @@ namespace TickTackk\DeveloperTools\XF\Repository;
 use TickTackk\DeveloperTools\XF\Repository\Exception\CodeEventNotFoundException;
 use XF\Entity\CodeEvent as CodeEventEntity;
 
+use function count;
+
 /**
  * Class CodeEvent
  * 
@@ -32,19 +34,19 @@ class CodeEvent extends XFCP_CodeEvent
         }
 
         $docBlock = "\t/**";
-        $descriptionParts = \array_filter(\explode('<p>', \str_ireplace([
+        $descriptionParts = array_filter(explode('<p>', str_ireplace([
             '<code>',
             '</code>'
         ], '', $codeEvent->description)));
 
         foreach ($descriptionParts AS $descriptionPartIndex => $descriptionPart)
         {
-            $descriptionPart = \utf8_trim($descriptionPart);
+            $descriptionPart = utf8_trim($descriptionPart);
             $descriptionParts[$descriptionPartIndex] = $descriptionPart;
 
-            if (\utf8_strlen($descriptionPart) > 4 && \utf8_substr($descriptionPart, \utf8_strlen($descriptionPart) - 4) === '</p>')
+            if (utf8_strlen($descriptionPart) > 4 && utf8_substr($descriptionPart, utf8_strlen($descriptionPart) - 4) === '</p>')
             {
-                $docBlock .= "\n\t * " . \wordwrap(\strip_tags(\utf8_rtrim($descriptionPart, '</p>')), 85, "\n\t * ") . \PHP_EOL;
+                $docBlock .= "\n\t * " . wordwrap(strip_tags(utf8_rtrim($descriptionPart, '</p>')), 85, "\n\t * ") . PHP_EOL;
                 unset($descriptionParts[$descriptionPartIndex]);
             }
         }
@@ -61,38 +63,38 @@ class CodeEvent extends XFCP_CodeEvent
                 'description' => html_entity_decode($parameterDescription)
             ];
 
-            $typeLength = \utf8_strlen($parameterType);
+            $typeLength = utf8_strlen($parameterType);
             if ($typeLength > $typePad)
             {
                 $typePad = $typeLength;
             }
 
-            $nameLength = \utf8_strlen($parameterName);
+            $nameLength = utf8_strlen($parameterName);
             if ($nameLength > $namePad)
             {
                 $namePad = $nameLength;
             }
         };
 
-        \preg_match('#<p>Arguments:<\/p>.<ol>(.*?)<\/ol>#si', $codeEvent->description, $matches);
+        preg_match('#<p>Arguments:<\/p>.<ol>(.*?)<\/ol>#si', $codeEvent->description, $matches);
 
-        $parametersBlock = \utf8_trim($matches[1] ?? ''); // trim LOL
-        $parameterLiCollection = \array_map(function ($li)
+        $parametersBlock = utf8_trim($matches[1] ?? ''); // trim LOL
+        $parameterLiCollection = array_map(function ($li)
         {
-            return \utf8_substr(\utf8_rtrim(\utf8_ltrim(\utf8_trim($li), '<li'), '</li>'), 1);
-        }, \explode(PHP_EOL, $parametersBlock)); // get cleaned <li> tags
+            return utf8_substr(utf8_rtrim(utf8_ltrim(utf8_trim($li), '<li'), '</li>'), 1);
+        }, explode(PHP_EOL, $parametersBlock)); // get cleaned <li> tags
 
         foreach ($parameterLiCollection AS $parameterLi)
         {
-            $parameterArr = \explode(' - ', $parameterLi);
-            if (\count($parameterArr) === 2)
+            $parameterArr = explode(' - ', $parameterLi);
+            if (count($parameterArr) === 2)
             {
                 $parameterDescription = $parameterArr[1];
 
-                if (\utf8_strpos($parameterArr[0], '<em>') === 0)
+                if (utf8_strpos($parameterArr[0], '<em>') === 0)
                 {
-                    [$parameterType, $parameterName] = \explode('</em>', $parameterArr[0]);
-                    $parameterType = \utf8_ltrim($parameterType, '<em>');
+                    [$parameterType, $parameterName] = explode('</em>', $parameterArr[0]);
+                    $parameterType = utf8_ltrim($parameterType, '<em>');
                 }
                 else
                 {
@@ -100,10 +102,10 @@ class CodeEvent extends XFCP_CodeEvent
                     $parameterName = $parameterArr[0];
                 }
 
-                $parameterType = \strip_tags($parameterType);
-                $parameterName = \strip_tags($parameterName);
+                $parameterType = strip_tags($parameterType);
+                $parameterName = strip_tags($parameterName);
 
-                $parameterName = \utf8_trim($parameterName);
+                $parameterName = utf8_trim($parameterName);
                 $addParameter($parameterName, $parameterDescription, $parameterType);
             }
         }
@@ -127,18 +129,18 @@ class CodeEvent extends XFCP_CodeEvent
                 $callbackSignature .= " {$name},";
             }
 
-            $type = \str_pad($type, $typePad);
-            $name = \str_pad($name, $namePad);
+            $type = str_pad($type, $typePad);
+            $name = str_pad($name, $namePad);
 
             $linePrefix = "\n\t * @param {$type} {$name}";
 
-            $description = \wordwrap($parameter['description'], 85, "\n\t" . \str_repeat(' ', \utf8_strlen($linePrefix) - 1));
+            $description = wordwrap($parameter['description'], 85, "\n\t" . str_repeat(' ', utf8_strlen($linePrefix) - 1));
 
             $docBlock .= $linePrefix . " {$description}";
         }
 
-        $callbackSignature = \rtrim(\utf8_trim($callbackSignature), ',');
+        $callbackSignature = rtrim(utf8_trim($callbackSignature), ',');
 
-        return \utf8_rtrim($docBlock, \PHP_EOL) . \PHP_EOL . "\t */";
+        return utf8_rtrim($docBlock, PHP_EOL) . PHP_EOL . "\t */";
     }
 }